Add files via upload
In a machine learning project aimed at predicting whether breast cancer is malignant or benign based on a large dataset, the primary goal is to develop a robust classification model. This project involves preprocessing and analyzing an extensive dataset of breast cancer cases, including various clinical and pathological features. By training the model on historical data, it can learn to identify patterns and relationships within the data. Once trained, the model can then be used to classify new breast cancer cases, providing valuable insights to healthcare professionals and aiding in early diagnosis and treatment decisions. The success of this project not only hinges on the model's accuracy but also on its potential to improve patient outcomes and contribute to advancements in medical diagnostics.